basic SQL defaults
