+++ flake.nix
@@ -23,7 +23,7 @@
         packages.default = pkgs.callPackage ./default.nix { };
         checks.default = pkgs.callPackage ./checks.nix { };
 
-        # devShells for people who are lazy with env setup
+        # devShells for people who are also lazy with env setup
         devShells.default = pkgs.mkShell {
           name = "merise_dot";
           inputsFrom = [ self.packages.${system}.default ];
